@@ -7,7 +7,7 @@ void main() {
 }
 
 class MyApp extends StatefulWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   State<MyApp> createState() => _MyAppState();
@@ -17,8 +17,9 @@ class _MyAppState extends State<MyApp> {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-        debugShowCheckedModeBanner: false,
-        routes: Routes.routes(context),
-        home: const FirstScreen());
+      debugShowCheckedModeBanner: false,
+      routes: Routes.routes(context),
+      home: const FirstScreen(),
+    );
   }
 }
