@@ -4,7 +4,7 @@ import 'package:navigator_plus/navigate.dart';
 import 'package:navigator_plus/result.dart';
 
 class SecondScreen extends StatefulWidget {
-  const SecondScreen({Key? key}) : super(key: key);
+  const SecondScreen({super.key});
 
   @override
   State<SecondScreen> createState() => _SecondScreenState();
@@ -19,37 +19,35 @@ class _SecondScreenState extends State<SecondScreen> {
     super.didChangeDependencies();
     _data = NavigatorName.getArguments(context);
     Navigate.extrasData(
-        context: context,
-        data: _data,
-        onSuccess: (data) {
-          log("Name: ${data['name']}");
-          _name = data['name'];
-        },
-        onFailure: () {
-          NavigatorName.finish(context);
-        },
-        keys: ["name"]);
+      context: context,
+      data: _data,
+      onSuccess: (data) {
+        log("Name: ${data['name']}");
+        _name = data['name'];
+      },
+      onFailure: () {
+        NavigatorName.finish(context);
+      },
+      keys: ["name"],
+    );
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('Second Screen'),
-      ),
+      appBar: AppBar(title: const Text('Second Screen')),
       body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
             Text(_name),
-            const SizedBox(
-              height: 20,
-            ),
+            const SizedBox(height: 20),
             ElevatedButton(
               onPressed: () {
-                NavigatorName.finish(context,
-                    result:
-                        Result(resultOk: true, data: {'ownerName': 'Ronak'}));
+                NavigatorName.finish(
+                  context,
+                  result: Result(resultOk: true, data: {'ownerName': 'Ronak'}),
+                );
               },
               child: const Text("Back to First Screen"),
             ),
