@@ -2,5 +2,4 @@ package com.datadirr.navigator_plus_example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity : FlutterActivity()
