@@ -2,9 +2,9 @@
 // of your plugin as a separate package, instead of inlining it in the same
 // package as the core of your plugin.
 // ignore: avoid_web_libraries_in_flutter
-import 'dart:html' as html show window;
 
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
+import 'package:web/web.dart' as web;
 
 import 'navigator_plus_platform_interface.dart';
 
@@ -20,7 +20,7 @@ class NavigatorPlusWeb extends NavigatorPlusPlatform {
   /// Returns a [String] containing the version of the platform.
   @override
   Future<String?> getPlatformVersion() async {
-    final version = html.window.navigator.userAgent;
+    final version = web.window.navigator.userAgent;
     return version;
   }
 }
