@@ -6,6 +6,8 @@
 
 #include <cstring>
 
+#include "navigator_plus_plugin_private.h"
+
 #define NAVIGATOR_PLUS_PLUGIN(obj) \
   (G_TYPE_CHECK_INSTANCE_CAST((obj), navigator_plus_plugin_get_type(), \
                               NavigatorPlusPlugin))
@@ -25,16 +27,20 @@ static void navigator_plus_plugin_handle_method_call(
   const gchar* method = fl_method_call_get_name(method_call);
 
   if (strcmp(method, "getPlatformVersion") == 0) {
-    struct utsname uname_data = {};
-    uname(&uname_data);
-    g_autofree gchar *version = g_strdup_printf("Linux %s", uname_data.version);
-    g_autoptr(FlValue) result = fl_value_new_string(version);
-    response = FL_METHOD_RESPONSE(fl_method_success_response_new(result));
+    response = get_platform_version();
   } else {
     response = FL_METHOD_RESPONSE(fl_method_not_implemented_response_new());
   }
 
   fl_method_call_respond(method_call, response, nullptr);
+}
+
+FlMethodResponse* get_platform_version() {
+  struct utsname uname_data = {};
+  uname(&uname_data);
+  g_autofree gchar *version = g_strdup_printf("Linux %s", uname_data.version);
+  g_autoptr(FlValue) result = fl_value_new_string(version);
+  return FL_METHOD_RESPONSE(fl_method_success_response_new(result));
 }
 
 static void navigator_plus_plugin_dispose(GObject* object) {
